Add files via upload
I had a problem with the project. Whenever I pressed Go Live and played the code it would always come up with white screen. I checked the console and it said that there were a lot of problems with the code which were already done by Byjus. For example there was one problem that the console said which was on line 29 in sketch. There was nothing I had to do in sketch so I do not know what to do.